timer modulization
