+++ hypervisor/Makefile
@@ -216,7 +216,11 @@ HW_C_SRCS += arch/x86/notify.c
 HW_C_SRCS += arch/x86/vtd.c
 HW_C_SRCS += arch/x86/gdt.c
 HW_C_SRCS += arch/x86/irq.c
-HW_C_SRCS += arch/x86/timer.c
+HW_C_SRCS += arch/x86/tsc.c
+HW_C_SRCS += arch/x86/tsc_deadline_timer.c
+HW_C_SRCS += common/timer.c
+HW_C_SRCS += common/udelay.c
+HW_C_SRCS += common/cycles.c
 HW_C_SRCS += arch/x86/vmx.c
 HW_C_SRCS += arch/x86/cpu_state_tbl.c
 HW_C_SRCS += arch/x86/pm.c
