+++ hypervisor/arch/x86/guest/vcpuid.c
@@ -14,6 +14,7 @@
 #include <vmx.h>
 #include <sgx.h>
 #include <logmsg.h>
+#include <cycles.h>
 
 static inline const struct vcpuid_entry *local_find_vcpuid_entry(const struct acrn_vcpu *vcpu,
 					uint32_t leaf, uint32_t subleaf)
@@ -177,7 +178,7 @@ static void init_vcpuid_entry(uint32_t leaf, uint32_t subleaf,
 			cpuid_subleaf(leaf, subleaf, &entry->eax, &entry->ebx, &entry->ecx, &entry->edx);
 		} else {
 			/* Use the tsc to derive the emulated 0x16U cpuid. */
-			entry->eax = (uint32_t) (get_tsc_khz() / 1000U);
+			entry->eax = (uint32_t) (get_cpu_freq() / 1000U);
 			entry->ebx = entry->eax;
 			/* Bus frequency: hard coded to 100M */
 			entry->ecx = 100U;
@@ -230,7 +231,7 @@ static void init_vcpuid_entry(uint32_t leaf, uint32_t subleaf,
 	 * EBX, ECX, EDX: RESERVED (reserved fields are set to zero).
 	 */
 	case 0x40000010U:
-		entry->eax = get_tsc_khz();
+		entry->eax = get_cpu_freq();
 		entry->ebx = 0U;
 		entry->ecx = 0U;
 		entry->edx = 0U;
