+++ hypervisor/arch/x86/guest/vlapic.c
@@ -44,6 +44,7 @@
 #include <ept.h>
 #include <trace.h>
 #include <logmsg.h>
+#include <cycles.h>
 #include "vlapic_priv.h"
 
 #define VLAPIC_VERBOS 0
@@ -272,14 +273,14 @@ static void vlapic_reset_timer(struct acrn_vlapic *vlapic)
 	timer = &vlapic->vtimer.timer;
 	del_timer(timer);
 	timer->mode = TICK_MODE_ONESHOT;
-	timer->fire_tsc = 0UL;
+	timer->timeout = 0UL;
 	timer->period_in_cycle = 0UL;
 }
 
 static bool
 set_expiration(struct acrn_vlapic *vlapic)
 {
-	uint64_t now = rdtsc();
+	uint64_t now = get_cpu_cycles();
 	uint64_t delta;
 	struct vlapic_timer *vtimer;
 	struct hv_timer *timer;
@@ -299,7 +300,7 @@ set_expiration(struct acrn_vlapic *vlapic)
 		if (vlapic_lvtt_period(vlapic)) {
 			timer->period_in_cycle = delta;
 		}
-		timer->fire_tsc = now + delta;
+		timer->timeout = now + delta;
 		ret = true;
 	}
 	return ret;
@@ -321,7 +322,7 @@ static void vlapic_update_lvtt(struct acrn_vlapic *vlapic,
 		del_timer(timer);
 		timer->mode = (timer_mode == APIC_LVTT_TM_PERIODIC) ?
 				TICK_MODE_PERIODIC: TICK_MODE_ONESHOT;
-		timer->fire_tsc = 0UL;
+		timer->timeout = 0UL;
 		timer->period_in_cycle = 0UL;
 
 		vtimer->mode = timer_mode;
@@ -330,19 +331,19 @@ static void vlapic_update_lvtt(struct acrn_vlapic *vlapic,
 
 static uint32_t vlapic_get_ccr(const struct acrn_vlapic *vlapic)
 {
-	uint64_t now = rdtsc();
+	uint64_t now = get_cpu_cycles();
 	uint32_t remain_count = 0U;
 	const struct vlapic_timer *vtimer;
 
 	vtimer = &vlapic->vtimer;
 
 	if ((vtimer->tmicr != 0U) && (!vlapic_lvtt_tsc_deadline(vlapic))) {
-		uint64_t fire_tsc = vtimer->timer.fire_tsc;
+		uint64_t deadline = vtimer->timer.timeout;
 
-		if (now < fire_tsc) {
+		if (now < deadline) {
 			uint32_t divisor_shift = vtimer->divisor_shift;
 			uint64_t shifted_delta =
-				(fire_tsc - now) >> divisor_shift;
+				(deadline - now) >> divisor_shift;
 			remain_count = (uint32_t)shifted_delta;
 		}
 	}
@@ -375,7 +376,7 @@ static void vlapic_write_icrtmr(struct acrn_vlapic *vlapic)
 		del_timer(&vtimer->timer);
 		if (set_expiration(vlapic)) {
 			/* vlapic_init_timer has been called,
-			 * and timer->fire_tsc is not 0, here
+			 * and timer->timeout is not 0,here
 			 * add_timer should not return error
 			 */
 			(void)add_timer(&vtimer->timer);
@@ -402,7 +403,7 @@ uint64_t vlapic_get_tsc_deadline_msr(const struct acrn_vlapic *vlapic)
 	} else if (!vlapic_lvtt_tsc_deadline(vlapic)) {
 		ret = 0UL;
 	} else {
-		ret = (vlapic->vtimer.timer.fire_tsc == 0UL) ? 0UL :
+		ret = (vlapic->vtimer.timer.timeout == 0UL) ? 0UL :
 			vcpu_get_guest_msr(vcpu, MSR_IA32_TSC_DEADLINE);
 	}
 
@@ -441,14 +442,14 @@ void vlapic_set_tsc_deadline_msr(struct acrn_vlapic *vlapic, uint64_t val_arg)
 		if (val != 0UL) {
 			/* transfer guest tsc to host tsc */
 			val -= exec_vmread64(VMX_TSC_OFFSET_FULL);
-			timer->fire_tsc = val;
+			timer->timeout = val;
 			/* vlapic_init_timer has been called,
-			 * and timer->fire_tsc is not 0,here
+			 * and timer->timeout is not 0, here
 			 * add_timer should not return error
 			 */
 			(void)add_timer(timer);
 		} else {
-			timer->fire_tsc = 0UL;
+			timer->timeout = 0UL;
 		}
 	} else {
 		/* No action required */
@@ -1374,7 +1375,7 @@ vlapic_write_svr(struct acrn_vlapic *vlapic)
 			if (vlapic_lvtt_period(vlapic)) {
 				if (set_expiration(vlapic)) {
 					/* vlapic_init_timer has been called,
-					 * and timer->fire_tsc is not 0,here
+					 * and timer->timeout is not 0,here
 					 *  add_timer should not return error
 					 */
 					(void)add_timer(&vlapic->vtimer.timer);
@@ -2038,7 +2039,7 @@ static void vlapic_timer_expired(void *data)
 	}
 
 	if (!vlapic_lvtt_period(vlapic)) {
-		vlapic->vtimer.timer.fire_tsc = 0UL;
+		vlapic->vtimer.timer.timeout = 0UL;
 	}
 }
