+++ hypervisor/arch/x86/pm.c
@@ -17,6 +17,8 @@
 #include <ioapic.h>
 #include <vtd.h>
 #include <lapic.h>
+#include <udelay.h>
+#include <cycles.h>
 
 struct cpu_context cpu_ctx;
 
@@ -163,7 +165,7 @@ void shutdown_system(void)
 
 static void suspend_tsc(__unused void *data)
 {
-	per_cpu(tsc_suspend, get_pcpu_id()) = rdtsc();
+	per_cpu(tsc_suspend, get_pcpu_id()) = get_cpu_cycles();
 }
 
 static void resume_tsc(__unused void *data)
