+++ hypervisor/common/ptdev.c
@@ -12,6 +12,7 @@
 #include <irq.h>
 #include <logmsg.h>
 #include <vtd.h>
+#include <cycles.h>
 
 #define PTIRQ_ENTRY_HASHBITS	9U
 #define PTIRQ_ENTRY_HASHSIZE	(1U << PTIRQ_ENTRY_HASHBITS)
@@ -176,10 +177,10 @@ static void ptirq_interrupt_handler(__unused uint32_t irq, void *data)
 			if (timer_is_started(&entry->intr_delay_timer)) {
 				to_enqueue = false;
 			} else {
-				entry->intr_delay_timer.fire_tsc = rdtsc() + entry->vm->intr_inject_delay_delta;
+				entry->intr_delay_timer.timeout = get_cpu_cycles() + entry->vm->intr_inject_delay_delta;
 			}
 		} else {
-			entry->intr_delay_timer.fire_tsc = 0UL;
+			entry->intr_delay_timer.timeout = 0UL;
 		}
 	}
