+++ hypervisor/debug/logmsg.c
@@ -11,6 +11,7 @@
 #include <per_cpu.h>
 #include <npk_log.h>
 #include <logmsg.h>
+#include <cycles.h>
 
 /* buf size should be identical to the size in hvlog option, which is
  * transfered to SOS:
@@ -53,10 +54,10 @@ void do_logmsg(uint32_t severity, const char *fmt, ...)
 	}
 
 	/* Get time-stamp value */
-	timestamp = rdtsc();
+	timestamp = get_cpu_cycles();
 
 	/* Scale time-stamp appropriately */
-	timestamp = ticks_to_us(timestamp);
+	timestamp = cycles_to_us(timestamp);
 
 	/* Get CPU ID */
 	pcpu_id = get_pcpu_id();
