+++ hypervisor/include/arch/x86/tsc.h
@@ -1,0 +1,17 @@
+/*
+ * Copyright (C) 2018 Intel Corporation. All rights reserved.
+ *
+ * SPDX-License-Identifier: BSD-3-Clause
+ */
+
+#ifndef TSC_H
+#define TSC_H
+
+/**
+ * @brief Calibrate tsc.
+ *
+ * @return None
+ */
+void calibrate_tsc(void);
+
+#endif
