+++ hypervisor/include/common/timer.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2018 Intel Corporation. All rights reserved.
+ * Copyright (C) 2020 Intel Corporation. All rights reserved.
  *
  * SPDX-License-Identifier: BSD-3-Clause
  */
@@ -8,6 +8,7 @@
 #define TIMER_H
 
 #include <list.h>
+#include <cycles.h>
 
 /**
  * @brief Timer
@@ -39,45 +40,11 @@ struct per_cpu_timers {
 struct hv_timer {
 	struct list_head node;		/**< link all timers */
 	enum tick_mode mode;		/**< timer mode: one-shot or periodic */
-	uint64_t fire_tsc;		/**< tsc deadline to interrupt */
+	uint64_t timeout;		    /**< tsc deadline to interrupt */
 	uint64_t period_in_cycle;	/**< period of the periodic timer in unit of TSC cycles */
 	timer_handle_t func;		/**< callback if time reached */
 	void *priv_data;		/**< func private data */
 };
-
-/* External Interfaces */
-
-#define CYCLES_PER_MS	us_to_ticks(1000U)
-
-void udelay(uint32_t us);
-
-/**
- * @brief convert us to ticks.
- *
- * @return ticks
- */
-uint64_t us_to_ticks(uint32_t us);
-
-/**
- * @brief convert ticks to us.
- *
- * @return microsecond
- */
-uint64_t ticks_to_us(uint64_t ticks);
-
-/**
- * @brief convert ticks to ms.
- *
- * @return millisecond
- */
-uint64_t ticks_to_ms(uint64_t ticks);
-
-/**
- * @brief read tsc.
- *
- * @return tsc value
- */
-uint64_t rdtsc(void);
 
 /**
  * @brief Initialize a timer structure.
@@ -85,7 +52,7 @@ uint64_t rdtsc(void);
  * @param[in] timer Pointer to timer.
  * @param[in] func irq callback if time reached.
  * @param[in] priv_data func private data.
- * @param[in] fire_tsc tsc deadline to interrupt.
+ * @param[in] timeout deadline to interrupt.
  * @param[in] mode timer mode.
  * @param[in] period_in_cycle period of the periodic timer in unit of TSC cycles.
  *
@@ -96,12 +63,12 @@ uint64_t rdtsc(void);
  */
 static inline void initialize_timer(struct hv_timer *timer,
 				timer_handle_t func, void *priv_data,
-				uint64_t fire_tsc, int32_t mode, uint64_t period_in_cycle)
+				uint64_t timeout, int32_t mode, uint64_t period_in_cycle)
 {
 	if (timer != NULL) {
 		timer->func = func;
 		timer->priv_data = priv_data;
-		timer->fire_tsc = fire_tsc;
+		timer->timeout = timeout;
 		timer->mode = mode;
 		timer->period_in_cycle = period_in_cycle;
 		INIT_LIST_HEAD(&timer->node);
@@ -117,7 +84,7 @@ static inline void initialize_timer(struct hv_timer *timer,
  */
 static inline bool timer_expired(const struct hv_timer *timer)
 {
-	return ((timer->fire_tsc == 0UL) || (rdtsc() >= timer->fire_tsc));
+	return ((timer->timeout == 0UL) || (get_cpu_cycles() >= timer->timeout));
 }
 
 /**
@@ -161,20 +128,6 @@ void del_timer(struct hv_timer *timer);
  * @return None
  */
 void timer_init(void);
-
-/**
- * @brief Calibrate tsc.
- *
- * @return None
- */
-void calibrate_tsc(void);
-
-/**
- * @brief  Get tsc.
- *
- * @return tsc(KHz)
- */
-uint32_t get_tsc_khz(void);
 
 /**
  * @}
