+++ hypervisor/include/common/udelay.h
@@ -1,0 +1,14 @@
+/*
+ * Copyright (C) 2018 Intel Corporation. All rights reserved.
+ *
+ * SPDX-License-Identifier: BSD-3-Clause
+ */
+
+#ifndef DELAY_H
+#define DELAY_H
+
+#include <types.h>
+
+void udelay(uint32_t us);
+
+#endif
